Implemented Java device stack bindings and moved some control point and core files into a better package structure.

+++ Upnp/Public/Java/ArgumentUint.h
@@ -10,10 +10,10 @@ extern "C" {
 /*
  * Class:     openhome_net_controlpoint_ArgumentUint
  * Method:    ActionArgumentCreateUintInput
- * Signature: (JI)J
+ * Signature: (JJ)J
  */
 JNIEXPORT jlong JNICALL Java_openhome_net_controlpoint_ArgumentUint_ActionArgumentCreateUintInput
-  (JNIEnv *, jobject, jlong, jint);
+  (JNIEnv *, jobject, jlong, jlong);
 
 /*
  * Class:     openhome_net_controlpoint_ArgumentUint
